Handling error code for defer functions (#545)
related to #124

because of the metrics we have to run the deferred function even if an error occurred during main processing. Until now, we used to call `os.Exit(1)` in main if any error occurred. Unfortunately, `os.exit` skips the deferred function call.

The solution is implementation. In case of an error we call `return` instead of `os.Exit`, and `os.Exit` itself is a deferred function.

goimports made some changes in terratest

Signed-off-by: kuritka <kuritka@gmail.com>