[Fix] busybox HitTestApp (#555)
Unfortunately I ran the crashed upgrade tests and lost the logs but #553 (upgrade tests) crashed on hitting busybox returning wget error.

What happened?
In our abstraction we use start test app which changes the replicaset to 1 instance of the subinfo. The application is Healthy , localtargets are in the correct state but the application itself is still booting. If I make an HTTP request while the application is booting, I get 503 instead of the expected json.

This fix polls the podinfo for 3 seconds while the application is booting. It also prints response.Body if the response is not the expected json.

Here is an example where polling was successfully used https://github.com/k8gb-io/k8gb/runs/3281617073?check_suite_focus=true#step:6:2492

Signed-off-by: kuritka <kuritka@gmail.com>