Fix #154: s/edgeDNSServer/edgeDNSServers/g
As described in the Issue #154 this adds the ability to fall-back to another edge dns server from the list. These are passed to the operator as env variable `EDGE_DNS_SERVERS` (was singular before) and the value is a comma-separted list of ips/hostnames followed optionally with port number. example: `1.1.1.1:11, 2.2.2.2,hostname` (spaces between entries are supported).

Then function `Dig` in `dns.go` has been modified to work with this list (variadic function) and also `Exchange` function in the same file was tweaked to support it.

This change doesn't break the backward compatibility, because both `EDGE_DNS_{SERVER,PORT}` still works, but user is notified in the operator logs that these two were deprecated.

Some old tests needed to be modified in order to pass and I've added also these new tests:
(`in controllers/depresolver/depresolver_test.go`)
-  TestResolveConfigWithEmptyEdgeDNSServersKey
-  TestResolveConfigWithMalformedEdgeDNSServersKey
-  TestResolveConfigWithTwoEdgeDnsServers
-  TestResolveConfigWithMultipleEdgeDnsServers1
-  TestResolveConfigWithMultipleEdgeDnsServers2
-  TestResolveConfigWithMultipleEdgeDnsServersMalformed1
-  TestResolveConfigWithMultipleEdgeDnsServersMalformed2
-  TestResolveConfigWithInvalidEdgeDnsServersValue
-  TestResolveConfigWithMultipleEdgeDnsServersLocalhostNotFirst

(in `controllers/gslb_controller_test.go`)
- TestReturnsExternalRecordsUsingFailoverStrategyAndFallbackDNSserver

(in `controllers/internal/utils/dns_test.go`)
-  TestOneValidEdgeDNSInTheList
-  TestNoValidEdgeDNSInTheList
-  TestEmptyEdgeDNSInTheList
-  TestMultipleValidEdgeDNSInTheList
-  TestEmptyDNSList
-  TestValidDigFQDNWithDot

Signed-off-by: Jirka Kremser <jiri.kremser@gmail.com>