DNS package test coverage (3/3)
This is a follow-up PR to #630 (1/3), do not review until #630 is merged.

Thanks to the ntroducing mocks I can test what is going on inside the large functions in the DNS package.
I have written several tests that create/update the DNS record and assert values at runtime (not just asserting the error the function returns).
I managed to achieve 62% coverage. The uncovered are mostly constructor functions, single-line functions that call the assistant package directly (this is more about testing the assistant package).

Signed-off-by: kuritka <kuritka@gmail.com>