Bind env variables with ENV-BINDER, clean tests
Thanks to ENV-BINDER we have moved the responsibility for bindings of environment variables to an external, well tested package.

ENV-BINDER works with unexported fields, so we can bind the values of some fields and calculate a new value from them in the calculation phase.
An example is enum, where env_variable has the value "info" but the internal state of the structure has the value 1. (I have also created a
PR in [kelsyhightower/envconfig](https://github.com/kelseyhightower/envconfig/pull/198), which solves the problem with private fields).

Thanks to the delegation of responsibility I could delete about a third of the test without changing the depressolver
 coverage, which increased to 95% with the changes.

Signed-off-by: kuritka <kuritka@gmail.com>