Fix upgrade-candidate
`upgrade-candidate` - the target for deploying the current version of
k8gb to local k3d clusters tried to deploy a local image that did not exist.
Here is the fix.

The go install command behaves almost identically to go build, but instead
of leaving the executable in the current directory, or a directory specified
by the -o flag, it places the executable into the $GOBIN directory.
goreleaser can be found and removed from there any time.

Signed-off-by: kuritka <kuritka@gmail.com>