gracefully finishing deferred calls 2
We recently solved how to exit k8gb and make sure all deferred calls are done. The solution was to run a defer at the very end of the program, which decided whether to run os.Exit(0) or os.Exit(1).

Although this solution works, it is still somewhat difficult to understand and quite error prone from the point of view that from the very start of the program, we have to maintain the state of how the program exits.

The following refactoring uses a different pattern - which is stateless and will be accepted by the kubebuilder when the PR is created.

```go
package main

func main() {
	if err := run(); err != nil {
		os.Exit(1)
	}
}

func run() error {

	// Current main() code with logs <-- logging, initializing etc.. is here

	return nil
}

```

Signed-off-by: kuritka <kuritka@gmail.com>