Always use the bash as the default shell
If the SHELL variable is not specified in the makefile, the env var is used if it exist. If not,
then the /bin/sh alias will be used. This is unpredictable because each OS can have different default
shell implementaion. For instance on Ubuntu it is dash and it behaves differently than bash.

Signed-off-by: Jirka Kremser <jiri.kremser@gmail.com>