Bump operator SDK to v1.18.1
I have checked / made the necessary changes in our operatorSDK for the newly released versions:
 - [1.17.0](https://sdk.operatorframework.io/docs/upgrading-sdk-version/v1.17.0/)
 - [1.18.0](https://sdk.operatorframework.io/docs/upgrading-sdk-version/v1.18.0/)
 - [1.18.1](https://sdk.operatorframework.io/docs/upgrading-sdk-version/v1.18.1/)

Modules: we had bumped versions already due to CVE exposed by dependabot. So we don't need to upgrade the modules now.
Other: it turned out that the only thing we need to update on our project is the controller-gen.

Signed-off-by: kuritka <kuritka@gmail.com>