Configuration weight (1/4)
Related to #50
The weight (optional field) configuration goes into the GSLB strategy. This PR only aims
to load and validate the weight configuration, which is specified in percentages.

```yaml
  strategy:
    type: roundRobin # Use a round robin load balancing strategy, when deciding which downstream clusters to route clients too
    weight: 20%
```
gslb.spec.strategy.weight is not allowed for failover and geoip strategies. If it is not specified in roundRobin,
then it is not a weight RoundRobin but classic one.

Signed-off-by: kuritka <kuritka@gmail.com>

move RoundRobinStrategy, FailoverStrategy, GeoStrategy constants into depresolver package