Refactoring SetupManager (#948)
* Refactoring SetupManager
related to #932

The setup manager is quite complex. We are currently solving a bug related to the fact
that changes in ingress also change the GSLB and can cause cycles #932. I am doing the
necessary refactoring (code is better structured, but the functionality remains the same),
to make it easier to make the necessary changes..

 - split SetupWithManager and Reconciliation into two files
 - encapsulate createGslbFromIngress
 - simplify ingressHandler
 - extracting parse strategy func

Signed-off-by: kuritka <kuritka@gmail.com>

* Update controllers/gslb_controller_setup.go

Co-authored-by: Jirka Kremser <535866+jkremser@users.noreply.github.com>
Signed-off-by: MichalK <kuritka@gmail.com>

Signed-off-by: kuritka <kuritka@gmail.com>
Signed-off-by: MichalK <kuritka@gmail.com>
Co-authored-by: Jirka Kremser <535866+jkremser@users.noreply.github.com>