Update references, GO1.19 (#1020)
I'm upgrading the dependency version because of dependabot. Unfortunately the change
is huge and all liners started to report different types of errors. That's why I'm upgrading
the liners, which in turn requires upgrading the GO version.
With all these changes I had to make changes in the code and linters because of deprecated
libraries and linters. I also had to make changes to our build pipelines.

Signed-off-by: kuritka <kuritka@gmail.com>

Signed-off-by: kuritka <kuritka@gmail.com>