Bump golangci to v1.59.1
I made the changes that were required to pass the linter. I removed the internal `directory` which hasn't made much sense for a long time. I've modified the GitHub action

the new linter has depguard as the main change, which is a guard for imported packages. these must be explicitly enabled/disabled in the configuration.
- https://golangci-lint.run/usage/linters/#depguard
- https://github.com/OpenPeeDeeP/depguard

Signed-off-by: Michal Kuritka <kuritka@gmail.com>