Fix ingress Annotations replacement loop
Annotation implementation change, terratests covered, test change.

When I create GSLB - not k8gb annotations are not copied to ingress. Annotations that are directly related to k8gb are now created in ingress.

When I change an annotation in ingress, the annotation is not copied to the GSLB

If I change an annotation that is related to k8gb in ingress, the GSLB as ource of truth returns the annotation

When I change an annotation in ingress that is not related to k8gb, reconciliation is not triggered and the annotation remains in ingress and is ignored by k8gb

When I create a GSLB from an ingress, the annotations are not copied to the GSLB.

Signed-off-by: Michal Kuritka <kuritka@gmail.com>