Fix failover strategy with more than 3 clusters
The failover strategy is not returning the correct targets if the DNS query hits a non-primary region

How to reproduce:

Create a 3 cluster setup that contains a GSLB with failover strategy with `eu` as primary cluster. The podinfo app is up on all clusters.
```
K8GB_LOCAL_VERSION=test CLUSTERS_NUMBER=3 make deploy-full-local-setup
```

Retrieve the local targets to see which IP addresses are exposed by each cluster (`eu`, `us` and `cz` respectively):
```
➜  ~ dig +short -p 5053 @localhost localtargets-failover.cloud.example.com
172.19.0.6
172.19.0.7
➜  ~ dig +short -p 5054 @localhost localtargets-failover.cloud.example.com
172.19.0.10
172.19.0.11
➜  ~ dig +short -p 5055 @localhost localtargets-failover.cloud.example.com
172.19.0.14
172.19.0.15
```

Query the domain `failover.cloud.example` on each cluster:
```
➜  ~ dig +short -p 5053 @localhost failover.cloud.example.com
172.19.0.6
172.19.0.7
➜  ~ dig +short -p 5054 @localhost failover.cloud.example.com
172.19.0.6
172.19.0.7
172.19.0.14
172.19.0.15
➜  ~ dig +short -p 5055 @localhost failover.cloud.example.com
172.19.0.6
172.19.0.7
172.19.0.10
172.19.0.11
```
We would expect all clusters to return 172.19.0.6 and 172.19.0.7, however the non-primary clusters return also the IP addresses of the other cluster:
* `eu` returns `eu`
* `us` returns `eu` and `cz`
* `cz` returns `eu` and `us`

This happens because a non-primary cluster returns the IP addresses of all other clusters, which is correct in a 2 cluster setup but not on a 3 cluster setup.

Fix

To fix the issue the IP addresses of the primary cluster must be returned if they exist. If the application is down on the primary cluster then all addresses of all clusters should be returned.

After the change:
```
➜  ~ dig +short -p 5053 @localhost failover.cloud.example.com
172.19.0.6
172.19.0.7
➜  ~ dig +short -p 5054 @localhost failover.cloud.example.com
172.19.0.6
172.19.0.7
➜  ~ dig +short -p 5055 @localhost failover.cloud.example.com
172.19.0.6
172.19.0.7
```

After scaling down podinfo on the primary cluster (`eu`):
```
➜  ~ k scale deploy frontend-podinfo --replicas=0 -n test-gslb --context k3d-test-gslb1
deployment.apps/frontend-podinfo scaled
```

```
➜  ~ dig +short -p 5053 @localhost failover.cloud.example.com
172.19.0.10
172.19.0.11
172.19.0.14
172.19.0.15
➜  ~ dig +short -p 5054 @localhost failover.cloud.example.com
172.19.0.10
172.19.0.11
172.19.0.14
172.19.0.15
➜  ~ dig +short -p 5055 @localhost failover.cloud.example.com
172.19.0.14
172.19.0.15
172.19.0.10
172.19.0.11
```

Signed-off-by: Andre Baptista Aguas <andre.aguas@protonmail.com>