fix chainsaw bug
When asserting the DNSEndpoint only check the length of the localtargets. This assertion is there to check if the status of the local cluster is correct but the "global targets" depend on the other cluster.

Also, increase the timeout to check the status of DNSEndpoint since it was reported they were not created early enough.