Add namespace label to terratests (#1846)
The reason for this PR is to simplify the cleaning process in terratests. If a test does not complete
for any reason (e.g., due to debugging), the namespace remains in the testing cluster. This can later
lead to confusion. Labeling namespaces significantly simplifies the cleaning of the test cluster.

Signed-off-by: Michal Kuritka <kuritka@gmail.com>