Remove finalization logic
When the GSLB was deleted, the DNS endpoint was also deleted, specifically the dnsendpoint in the ExternalDNS case or directly the DNS zone from Infoblox. But what is the point of DNS zone deletion in the case where the dnsendpoint is on k8gb? Killing the k8gb pod? Manual deletion of the dnsendpoint?

So, DNS zones are not finalized.