DNSZones setup on bootstrap
- dnsendpoint refactored to standalone package. We have two implementations
   - applicationDNSEndpoint
   - delegationZoneDNSEndpoint

Creating a delegation zone during bootstrap - the necessary IP addresses are obtained from coreDNSService
(if the type is LoadBalancer). If it is not this service must include an annotation (k8gb.io/coredns-ingress-ref)
that contains the ingress, which includes these addresses. The format is <namespace>/ingress.
Locally, we use init-ingress in the k8gb namespace.

- We added secondary reconciliation (CoreDNSReconciler) creates dns zones
    - on bootstrap
    - when k8gb-coredns service has changed
    - when the ingress specified in the annotation (k8gb.io/coredns-ingress-ref) in the k8gb-coredns service changes.
    - otherwise reconciliation waits.

- Added chainsaw test to check whether correct endpoint exists

I removed finalizers.From the GSLB perspective, application DNS endpoints are automatically removed thanks to the owner reference.
The previous implementation of the GSLB finalizer was incorrect. After removing the GSLB, the delegationZone DNS Endpoint
was also removed—and was subsequently created by another GSLB during reconciliation.

Now, finalization is no longer necessary; however, the interface still accounts for finalizers and within implementation is ready for them.
- finalizators are disabled and fixed. BEFORE: Finalizer of GSLBE removed zone delegation which was recreated back by another GSLB

I refactored assistant - now it returns only resource and IPs. We have implementations for k8gb-coredns and ingress mentioned above
- Create zone delegation refactored