DNSZones setup on bootstrap
If we talk about enterprise environments, it is easy to encounter a situation where we have thousands of GSLBs and a single DNS server. This server can be easily overwhelmed. Currently, each reconciliation (which occurs every 30 seconds by default) results in a read hit on the DNS server. With such a large number of GSLBs, this becomes a problem.

This PR moves the DNS zone configuration to the bootstrap process, eliminating the need to query the DNS on every reconciliation. Additionally, it fixes finalizers, preventing the recreation of a DNS zone each time a GSLB is deleted.

The solution works by introducing a second reconciliation process that only triggers on startup or when changes occur in the k8gb-coredns service or init-ingress.

If k8gb-coredns service is of type LoadBalancer, we simply read the IP addresses and create the DNS zones on the DNS server (including external-DNS endpoints in the k8gb namespace).
If LoadBalancer is not used, the k8gb-coredns service must contain the annotation k8gb.io/coredns-ingress-ref=<init-ingress-namespace/init-ingress-name>, from which k8gb will retrieve the IP addresses.
Lastly, the assistant has been refactored, and external DNS endpoints have been separated into an independent implementation (k8gbendpoint.DNSEndpoint interface, with implementations: delegationZoneDNSEndpoint and applicationDNSEndpoint). This improves code maintainability.

- dnsendpoint refactored to standalone package. We have two implementations
   - applicationDNSEndpoint
   - delegationZoneDNSEndpoint

Creating a delegation zone during bootstrap - the necessary IP addresses are obtained from coreDNSService
(if the type is LoadBalancer). If it is not this service must include an annotation (k8gb.io/coredns-ingress-ref)
that contains the ingress, which includes these addresses. The format is <namespace>/ingress.
Locally, we use init-ingress in the k8gb namespace.

- We added secondary reconciliation (CoreDNSReconciler) creates dns zones
    - on bootstrap
    - when k8gb-coredns service has changed
    - when the ingress specified in the annotation (k8gb.io/coredns-ingress-ref) in the k8gb-coredns service changes.
    - otherwise reconciliation waits.

- Added chainsaw test to check whether correct endpoint exists

I removed finalizers.From the GSLB perspective, application DNS endpoints are automatically removed thanks to the owner reference.
The previous implementation of the GSLB finalizer was incorrect. After removing the GSLB, the delegationZone DNS Endpoint
was also removed—and was subsequently created by another GSLB during reconciliation.

Now, finalization is no longer necessary; however, the interface still accounts for finalizers and within implementation is ready for them.
- finalizators are disabled and fixed. BEFORE: Finalizer of GSLBE removed zone delegation which was recreated back by another GSLB

I refactored assistant - now it returns only resource and IPs. We have implementations for k8gb-coredns and ingress mentioned above
- Create zone delegation refactored

- Replace Unit tests
- For multizones externalDNS v0.12.0 andhigher must be provided
- Chainsaw with multizones

Signed-off-by: Michal Kuritka <kuritka@gmail.com>