Fix status field requirements
When migrating from 0.12 I've found that updates to any gslb object was failing due to validation errors.

This happens when resourceRef is not set yet, object is saved with .spec.resourceRef but status.loadbalancer nor status.servers is not yet set
or calculated.

Lets relax a requirement to allow smoother transition to updated status structure

Signed-off-by: Dinar Valeev <k0da@opensuse.org>