Bootstrap refactored
Just like before, when coreDNS serviceType=LoadBalancer, the IP addresses from this service will be used for delegationDNSEndpoint.
The change is that if the serviceType is not LoadBalancer, the IP addresses of the first found ingress with the label k8gb.io/ip-source=true will be used instead.
Additionally, I refactored the CoreDNS controller setup and the assistant.

Service type  is determined by our Helm values.yaml file, specifically by the coredns.serviceType value.

Signed-off-by: Michal Kuritka <kuritka@gmail.com>