[breaking] deprecate dnsZone and edgeDnsZone chart values
PR #1845 introduced the possibility to configure multiple DNS zones. The intention was to be backwards compatible. However it forces uses to change the following helm values, as explained in #1858:
* k8gb.dnsZoneNegTTL -> k8gb.dnsZones[0].dnsZoneNegTTL
* k8gb.edgeDnsZone -> k8gb.dnsZones[0].zone
* k8gb.dnsZone -> k8gb.dnsZones[0].domain

During the community meeting of 02.04.2025 it was agreed to maintain the breaking change and clean up the deprecated values configuration.

Fixes #1858

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>