Waiting for ingress IP addresses in the local environment, fix logger.Interface()
If the bootstrap address does not exist, k8gb fails and the pod restarts. Unfortunately, the logger was used incorrectly — the Interface() function was used, which performs unmarshalling and caused panics. The logger has now been fixed.

Additionally, in the local environment, it is necessary to wait until the ingress is bound; otherwise, the k8gb pod keeps restarting during tests until the ingress is installed.

Signed-off-by: Michal Kuritka <kuritka@gmail.com>