Fix extdns/k8gb dns zone validation
We have a validation function that verifies if the DNS zones configured in k8gb and external DNS match, to prevent misconfigurations.

One of the conditions was checking the number of zones is the same in both configurations, however it is possible to have valid configurations where this is not the case. Take the example:
```
k8gb:
  dnsZones:
  - parentZone: example.com
    loadBalancedZone: cloud.example.com
  - parentZone: example.com
    loadBalancedZone: demo.example.com
```
In this case, the parentZones are the same, so we only need one entry in the external-dns configuration.

There is a workaround though. One can duplicate the entry in the external dns configuration without any side effects:
```
extdns:
  domainFilters:
  - example.com
  - example.com
```

To clean up the validation and avoid this redundancy the PR removes the faulty equality condition. It is enough to verify if each k8gb's parentZone is present in the external-dns's domainFilters, and vice-versa.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>