Remove unused route53 helm key (#1997)
With the migration of route53 to external dns in v0.15.0, the `route53` key in the helm chart is no longer used. Therefore, this PR removes it.

In addition, I noticed there was a documentation example of route53 configuration that was forgotten. It is now updated to reflect the new configuration.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>