Reload coredns if configuration changes (#2000)
Until now, coreDNS pods needed a restart to pick up configuration updates (e.g. adding or removing DNS zones)
By enabling the reload plugin, the configuration of the Corefile is automatically picked up upon changes. It is a hot reload, the pods do not restart. Also, any Corefile changes are picked up, not only the ones in the zone where the plugin is configured.

The reload interval and jitter are made configurable and set to the default values documented in https://coredns.io/plugins/reload/

Query coredns for a record in a zone not yet configured (e.g. podinfo.aws.k8gb.website), which fails.
Add the zone "aws.k8gb.website" to the core file.
Wait for the reload
Query coredns for the same record, which now resolves correctly.

The reload is also visible in the logs:
```
[INFO] Reloading
[INFO] plugin/k8s_crd: Filter: [k8gb.absa.oss/dnstype=local]
[INFO] plugin/k8s_crd: negTTL: 30
[INFO] plugin/k8s_crd: loadbalance: [weight]
[INFO] plugin/k8s_crd: Running 'inCluster' kube controller
[INFO] plugin/k8s_crd: Filter: [k8gb.absa.oss/dnstype=local]
[INFO] plugin/k8s_crd: negTTL: 30
[INFO] plugin/k8s_crd: loadbalance: [weight]
[INFO] plugin/k8s_crd: Running 'inCluster' kube controller
[INFO] plugin/reload: Running configuration SHA512 = a08ef6c491feee319ca8e583a1abb865c1efe853d160051bba7ade67805ac519761242fa2e2798378b322f48692bac6a619bc8a6efb90e8ae7b934f27709549e
[INFO] Reloading complete
```

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>