Refactor namespace initialization in Chainsaw
We had different initializations for istio and non-istio setups because of the istio-inject label on the namespace.

This PR consolidates the initialization in a single step template that sets the label dynamically.

The validation of the Ingress resource is also removed to allow this consolitation. This can be done because a valid DNSEndpoint guarantees a valid Ingress.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>