remove failover playground terratest (#2031)
* remove failover playground terratest

The terratest setup is quite complex, which resulted in us having to give up on test parallelism. This increased the testing time, so we are moving away from it to Chainsaw.

The failover playground test was already implemented in Chainsaw a few months ago and it has been running well. We can now remove this test from the Terratest setup and some some precious e2e testing minutes.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>

* pin helm version

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>

---------

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>