replicate basic_app_test in chainsaw
Createst the exact same testing logic in Chainsaw as the test: terratest/test/k8gb_basic_app_test.go

The test applies a GSLB resource with an embedded Ingress. The GSLB contains 3 ingress rules:
* one app is healthy
* one app is unhealthy
* one app does not exist

Then, an unhealthy and a healthy app are created.

The status of GSLB is checked to verify it the service health of the apps is Healthy, Unhealthy and NotFound, respectively.

The DNSEndpoint is checked to verify if it contains the label `k8gb.absa.oss/dnstype` and an owner reference.

2 broken GSLB resources are applied to verify if they are rejected by the API

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>