replicate basic_app_test in chainsaw (#2038)
Creates the exact same testing logic in Chainsaw as the test: [terratest/test/k8gb_basic_app_test.go](https://github.com/k8gb-io/k8gb/blob/master/terratest/test/k8gb_basic_app_test.go)
    
The test applies a GSLB resource with an embedded Ingress. It then verifies if the created Ingress resource is correctly created picked up by k8gb. Therefore, the test was renamed to embedded ingress creation

### Test description

The GSLB contains 3 ingress rules:
* one app is healthy
* one app is unhealthy
* one app does not exist

Then, an unhealthy and a healthy app are created, and a couple of status checks are done. The status of GSLB is checked to verify it the service health of the apps is Healthy, Unhealthy and NotFound, respectively. The DNSEndpoint is checked to verify if it contains the label `k8gb.absa.oss/dnstype` and an owner reference.

Finally, 2 broken GSLB resources are applied to verify if they are rejected by the API
    
Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>