Fix Prometheus and Grafana exposure (#2109)
In the local setup, Prometheus and Grafana were exposed as nodePort services. In my machine it was not always possible to reach the servers, so I now expose them as loadbalancer services, similar to how coreDNS is exposed.

---

In addition, the connection between `grafana` and the `prometheus` servers (one in each cluster) was fixed. This connection was previously done via direct browser access. However, this feature is now deprecated:
```
Browser access mode in the Prometheus data source is no longer available. Switch to server access mode.
```
So we need to connect using `proxy` access (from the container). We can use the address `172.17.0.1` to have access to the host network.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>