Add support for GatewayAPI's HTTPRoute and GRPCRoute (#2114)
This is a first of a series of PRs implementing support for the GatewayAPI: #954

Gateway API quick intro to understand the CRDs: https://www.youtube.com/watch?v=6m6-VrQPXyY

This PR adds the Gateway API CRDs to the local setup, and adds support for [HTTPRoutes](https://gateway-api.sigs.k8s.io/api-types/httproute/) and [GRPCRoutes](https://gateway-api.sigs.k8s.io/api-types/grpcroute/).

For k8gb, the key components of any route resource are the `parentRefs` (linking to the Gateway resource where we can fetch the exposed IP addresses), the `backendRefs` (linking to the Services backends, from where we determine the health of the application), and the hostname of the application being exposed. Here is an example using HTTPRoute:
```
apiVersion: gateway.networking.k8s.io/v1
kind: HTTPRoute
metadata:
  name: gatewayapi-httproute
  namespace: test-gslb
spec:
  parentRefs:
  - name: gatewayapi-gateway
  hostnames:
  - gatewayapi-httproute.cloud.example.com
  rules:
  - backendRefs:
    - name: gatewayapi-service
      port: 8080
```

A Gateway looks as follows:
```
apiVersion: gateway.networking.k8s.io/v1
kind: Gateway
metadata:
  name: gatewayapi-gateway
  namespace: test-gslb
spec:
  gatewayClassName: gatewayapi-gateway
  listeners:
  - name: http
    protocol: HTTP
    port: 80
status:
  addresses:
  - type: IPAddress
    value: 10.0.0.1
```

The only slightly involved logic in this PR is in the `adapters.go` file. There, we define two interfaces: `routeSpec` and a `routeRuleSpec`. These interfaces are used to wrap the different route resources, since the fields we are interested in are defined in different structs for each Route type (GRPCRoute, HTTPRoute, ...). Our business logic can then use it to fetch `parentRefs`, `backendRefs` and `hostnames`, independently of the underlying route type.

The rest of the code is simply fetching the right field from the Custom Resources, similar to what we do with Ingress or VirtualService resources.

---

The planned course of action for follow up PRs are:
- support for TCPRoute, UDPRoute and TLSRoute
- E2E testing in chainsaw
- integration in the local setup